+++ src/main/java/com/finalProject/model/admin/coupon/CouponDTO.java
@@ -1,6 +1,4 @@
-package com.finalProject.model.admin;
-
-import java.sql.Timestamp;
+package com.finalProject.model.admin.coupon;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,15 +7,18 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 
+
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
 @Setter
 @ToString
 @Builder
-public class CouponPayDTO {
+public class CouponDTO {
 	private int coupon_no;
-	private String coupon_code;
-	private String member;
-	private Timestamp expire_date;
+	private String coupon_name;
+	private String coupon_dc_type;
+	private int coupon_dc_amount;
+	private float coupon_dc_rate;
+	private int coupon_use_days;
 }
