+++ src/main/java/com/finalProject/model/admin/coupon/CouponPayDTO.java
@@ -1,4 +1,6 @@
-package com.finalProject.model.admin;
+package com.finalProject.model.admin.coupon;
+
+import java.sql.Timestamp;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -7,18 +9,15 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 
-
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
 @Setter
 @ToString
 @Builder
-public class CouponDTO {
+public class CouponPayDTO {
 	private int coupon_no;
-	private String coupon_name;
-	private String coupon_dc_type;
-	private int coupon_dc_amount;
-	private float coupon_dc_rate;
-	private int coupon_use_days;
+	private String coupon_code;
+	private String member;
+	private Timestamp expire_date;
 }
