+++ src/main/java/com/finalProject/model/admin/coupon/PagingInfoNewDTO.java
@@ -1,4 +1,4 @@
-package com.finalProject.model.admin;
+package com.finalProject.model.admin.coupon;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
