+++ src/main/java/com/finalProject/service/admin/coupon/CouponServiceImpl.java
@@ -14,10 +14,10 @@ import javax.inject.Inject;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.finalProject.model.admin.CouponDTO;
-import com.finalProject.model.admin.CouponPayDTO;
-import com.finalProject.model.admin.PagingInfoNew;
-import com.finalProject.model.admin.PagingInfoNewDTO;
+import com.finalProject.model.admin.coupon.CouponDTO;
+import com.finalProject.model.admin.coupon.CouponPayDTO;
+import com.finalProject.model.admin.coupon.PagingInfoNew;
+import com.finalProject.model.admin.coupon.PagingInfoNewDTO;
 import com.finalProject.persistence.admin.coupon.CouponDAO;
 
 @Service
