+++ src/main/resources/mybatis-config.xml
@@ -6,13 +6,16 @@
 	<typeAliases>
 		<typeAlias type="com.finalProject.model.LoginDTO" alias="LoginDTO" />
 		<typeAlias type="com.finalProject.model.admin.notices.NoticeDTO" alias="NoticeDTO" />
-		<typeAlias type="com.finalProject.model.admin.MemberManageDTO" alias="MemberManageDTO" />
-		<typeAlias type="com.finalProject.model.admin.CouponDTO" alias="CouponDTO" />
-		<typeAlias type="com.finalProject.model.admin.CouponPayDTO" alias="CouponPayDTO" />
+		
 		<typeAlias type="com.finalProject.model.MemberDTO" alias="MemberDTO" />
 		<typeAlias type="com.finalProject.model.ProductDTO" alias="ProductDTO" />
 		
+		
 		<typeAlias type="com.finalProject.model.inquiry.InquiryDTO" alias="InquiryDTO" />
 		<typeAlias type="com.finalProject.model.inquiry.InquiryDetailDTO" alias="InquiryDetailDTO" />
+		
+		<typeAlias type="com.finalProject.model.admin.member.MemberManageDTO" alias="MemberManageDTO" />
+		<typeAlias type="com.finalProject.model.admin.coupon.CouponDTO" alias="CouponDTO" />
+		<typeAlias type="com.finalProject.model.admin.coupon.CouponPayDTO" alias="CouponPayDTO" />
 	</typeAliases>
 </configuration>
